Fix issues in common folder
